sst: move helper packages to internal/
Make them not importable by other projects.
