+++ pkg/sst/internal/isst/_types_amd64.go
