+++ pkg/sst/internal/isst/gen_types.sh
