+++ pkg/sst/internal/isst/types_amd64.go
