+++ pkg/sst/internal/mbox/mbox.go
@@ -19,7 +19,7 @@ package mbox
 import (
 	"fmt"
 
-	"github.com/intel/goresctrl/pkg/sst/isst"
+	"github.com/intel/goresctrl/pkg/sst/internal/isst"
 )
 
 // numClos is the number of CLOSes supported by SST-CP.
